requirements.txt
python-dotenv
Flask
WTForms
Flask-WTF
pymongo
email-validator
scikit-learn
python-docx
PyPDF2
spacy
Flask-Session
Werkzeug
passlib
argon2-cffi
Flask-Mail
pyotp
torch
transformers
pytorch_lightning
sentencepiece
tokenizers
huggingface-hub
jsonify
langchain-pinecone
langchain
langchain-google-genai
langchain-community
sentence-transformers
pinecone-client
pinecone[grpc]